Added anonymous sign in functionality
