+++ lib/models/user.dart
@@ -1,0 +1,4 @@
+class User {
+  final String uid;
+  User({this.uid});
+}
