+++ lib/screens/authentication/login.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:i_scanner/screens/authentication/sign_in.dart';
+
+class Authenticate extends StatefulWidget {
+  @override
+  _AuthenticateState createState() => _AuthenticateState();
+}
+
+class _AuthenticateState extends State<Authenticate> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: SignIn(),
+    );
+  }
+}
