+++ lib/screens/authentication/sign_in.dart
@@ -1,0 +1,67 @@
+import 'package:flutter/material.dart';
+import 'package:i_scanner/services/auth.dart';
+
+class SignIn extends StatefulWidget {
+  @override
+  _SignInState createState() => _SignInState();
+}
+
+class _SignInState extends State<SignIn> {
+  final AuthService _auth = AuthService();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.grey[200],
+      appBar: AppBar(
+          backgroundColor: Colors.redAccent,
+          elevation: 0.0,
+          title: Text('Sign in to IScanner')),
+      body: Container(
+        padding: EdgeInsets.symmetric(vertical: 20.0, horizontal: 50.0),
+        child: RaisedButton(
+          child: Text('Sign in anon'),
+          onPressed: () async {
+            dynamic result = await _auth.signInAnon();
+            if (result == null) {
+              print('error signing in');
+            } else {
+              print('sign in successful');
+              print(result.uid);
+            }
+          },
+        ),
+      ),
+    );
+  }
+}
+
+/*class _SignInState extends State<SignIn> {
+  final AuthService _auth = AuthService();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.grey[200],
+      appBar: AppBar(
+          backgroundColor: Colors.redAccent,
+          elevation: 0.0,
+          title: Text('Sign in to IScanner')),
+      body: Container(
+        padding: EdgeInsets.symmetric(vertical: 20.0, horizontal: 50.0),
+        child: RaisedButton(
+          child: Text('Sign in anon'),
+          onPressed: () async {
+            dynamic result = await _auth.signInAnon();
+            if (result == null) {
+              print('error signing in');
+            } else {
+              print('sign in successful');
+              print(result.uid);
+            }
+          },
+        ),
+      ),
+    );
+  }
+}*/
