+++ lib/screens/home/extras/doc_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
-import 'package:i_scanner/pages/scanner_page.dart';
-import 'package:i_scanner/pages/document.dart';
+import 'package:i_scanner/screens/home/pages/scanner_page.dart';
+import 'package:i_scanner/screens/home/pages/document.dart';
 
 // card for displaying current added documents (images)
 class DocCard extends StatefulWidget {
