+++ pubspec.yaml
@@ -33,6 +33,7 @@ dependencies:
   flutter_launcher_icons: ^0.7.5
   firebase_auth: ^0.16.1
   cloud_firestore: ^0.13.7
+  provider: ^4.3.0
